[core] add go_lib module set_cr_statuses
