+++ go_lib/hooks/set_cr_statuses/hook.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2022 Flant JSC
+Copyright 2023 Flant JSC
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -14,27 +14,57 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package hooks
+package set_cr_statuses
 
 import (
+	"encoding/json"
 	"fmt"
 	"time"
 
-	"github.com/clarketm/json"
-	"github.com/flant/addon-operator/pkg/module_manager/go_hook"
-	"github.com/flant/addon-operator/sdk"
-	"github.com/flant/shell-operator/pkg/kube/object_patch"
 	utils_checksum "github.com/flant/shell-operator/pkg/utils/checksum"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
+	"github.com/flant/addon-operator/pkg/module_manager/go_hook"
 )
 
-// hook for setting CR statuses
-var _ = sdk.RegisterFunc(&go_hook.HookConfig{
-	Queue:       "/modules/admission-policy-engine/security_policies",
-	OnAfterHelm: &go_hook.OrderedConfig{Order: 10},
-}, annotateSP)
+var SetObservedStatus = func(snapshot go_hook.FilterResult, filterFunc func(*unstructured.Unstructured) (go_hook.FilterResult, error)) func(obj *unstructured.Unstructured) (*unstructured.Unstructured, error) {
+	snBytes, _ := json.Marshal(snapshot)
+	checkSum := utils_checksum.CalculateChecksum(string(snBytes))
 
-var processedStatus = func(filterFunc func(*unstructured.Unstructured) (go_hook.FilterResult, error)) func(obj *unstructured.Unstructured) (*unstructured.Unstructured, error) {
+	return func(obj *unstructured.Unstructured) (*unstructured.Unstructured, error) {
+		objCopy := obj.DeepCopy()
+		filteredObj, err := filterFunc(objCopy)
+		if err != nil {
+			return nil, fmt.Errorf("cannot apply filterFunc to object: %v", err)
+		}
+
+		objBytes, err := json.Marshal(filteredObj)
+		if err != nil {
+			return nil, fmt.Errorf("cannot marshal filtered object: %v", err)
+		}
+
+		objCheckSum := utils_checksum.CalculateChecksum(string(objBytes))
+		if checkSum == objCheckSum {
+			processedCheckSum, found, err := unstructured.NestedString(objCopy.Object, "status", "deckhouse", "processed", "checkSum")
+			if err != nil {
+				return nil, fmt.Errorf("cannot get processed checksum status field: %v", err)
+			}
+
+			if !found || checkSum != processedCheckSum {
+				if err := unstructured.SetNestedField(objCopy.Object, "False", "status", "deckhouse", "synced"); err != nil {
+					return nil, fmt.Errorf("cannot set synced status field: %v", err)
+				}
+			}
+			if err := unstructured.SetNestedStringMap(objCopy.Object, map[string]string{"lastTimestamp": time.Now().Format(time.RFC3339), "checkSum": checkSum}, "status", "deckhouse", "observed"); err != nil {
+				return nil, fmt.Errorf("cannot set observed status field: %v", err)
+			}
+		} else {
+			return nil, fmt.Errorf("object has changed since last snapshot")
+		}
+		return objCopy, nil
+	}
+}
+
+var SetProcessedStatus = func(filterFunc func(*unstructured.Unstructured) (go_hook.FilterResult, error)) func(obj *unstructured.Unstructured) (*unstructured.Unstructured, error) {
 	return func(obj *unstructured.Unstructured) (*unstructured.Unstructured, error) {
 		objCopy := obj.DeepCopy()
 		filteredObj, err := filterFunc(objCopy)
@@ -69,19 +99,4 @@ var processedStatus = func(filterFunc func(*unstructured.Unstructured) (go_hook.
 		}
 		return objCopy, nil
 	}
-}
-
-func annotateSP(input *go_hook.HookInput) error {
-	securityPolicies := make([]securityPolicy, 0)
-
-	err := json.Unmarshal([]byte(input.Values.Get("admissionPolicyEngine.internal.securityPolicies").String()), &securityPolicies)
-	if err != nil {
-		return fmt.Errorf("cannot unmarshal values: %v", err)
-	}
-
-	for _, sp := range securityPolicies {
-		input.PatchCollector.Filter(processedStatus(filterSP), "deckhouse.io/v1alpha1", "securitypolicy", "", sp.Metadata.Name, object_patch.WithSubresource("/status"))
-	}
-
-	return nil
 }
